Software list page and inline listing in posts
