+++ src/_content/software/pyspread.md
@@ -1,6 +1,6 @@
 ---
-displayName: "*pyspread*"
+displayName: "pyspread"
 url: "https://pyspread.gitlab.io/"
 ---
 
-_pyspread_ is a non-traditional spreadsheet application that is based on and written in the programming language Python. The goal of *pyspread* is to be the most pythonic spreadsheet.
+_pyspread_ is a non-traditional spreadsheet application that is based on and written in the programming language Python. The goal of _pyspread_ is to be the most pythonic spreadsheet.
